Mainly fixed linker script issue.
* Fixed the linker command line so that the provided linker script is
used and found.
* Fixed the linker script (the one that was in the repository was too
old to support the current locm3)
* Synchronized more of the compiler and linker flags to the locm3
examples.

TODO:
* remove the libopencm3 submodule. (if we want to provide our own
libopencm3 version we have to make it a framework package, the
submodules are only meant for library build scripts).
* Clean up the buildscripts, I am still not sure where what should be
added to the commandline of each and every part of the build.
* Find a way to build the linker scripts inside the libopencm3 instead
of using our own copy of it.
* allow the user to override the linker script
* find a way to use an up to date version of libopencm3 even if it means
we have to provide our own framework.